modularization: extract gcov routines
